extracted content classes to separate file
